Various Updates
- Fixed error in LocalDB with Database missing close statements
@audreyx
- Added My Tasks. More research will need to be done for tabs. Could
not get to work, would just display nothing at all on application.
- Changed Task back to Parcelable, only to get still nothing passing.
@SirMorek we may need to just query database again, as cannot figure
out why this isn't working. Parcelable apparently is the way your
supposed to do it. Put comment in code in Task.java for link to
stackoverflow about it.
- Added MyTasksActivity, still needs to implement the Linear Layout and
get all current Users tasks. This will need LocalDB to be updated with
getAllTasks(someUserID) and then put that info into TaskListAdapter
which will output to the Linear List.